Add confirmation for deleting questions
[Fixes #56581034]

+++ questions/static/js/session-owner.js
@@ -57,12 +57,15 @@ $("#session-desc-edit").blur(function(){
 function sessionClicksOwner(){
 	$(".delete").off("click");
 	$(".delete").click(function(e){
-	    e.preventDefault();
-	    $(this).closest(".question").remove();
-	    $.ajax("/api/v1/question/"+$(this).attr("data-question")+"/",
-	        {
-	    	'type':"DELETE"
-	    });
+		del = this
+		e.preventDefault();
+		confirm("Are you sure you want to delete this question?", function(){
+		    $.ajax("/api/v1/question/"+$(del).attr("data-question")+"/",
+		        {
+		    	'type':"DELETE"
+		    });
+		    $(del).closest(".question").hide().addClass("deleted");
+		});
 	});
 	$(".star").off("click");
 	$(".star").click(function(e){
